--- CaveSoSe2025/Assets/PositionTransfer2.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d4eadb0af62789e48ae0d7291e227646
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
