+++ CaveSoSe2025/Assets/Scenes/SampleScene.unity
@@ -952,57 +952,6 @@ Transform:
   m_Children: []
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 78.1, y: 0.32, z: 0.77}
---- !u!1 &1394094059
-GameObject:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  serializedVersion: 6
-  m_Component:
-  - component: {fileID: 1394094061}
-  - component: {fileID: 1394094060}
-  m_Layer: 0
-  m_Name: PositionTransfer
-  m_TagString: Untagged
-  m_Icon: {fileID: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 0
---- !u!114 &1394094060
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1394094059}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 342e7afcbea331d4eb1a5836c584ad07, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  mirrorPlanePoint: {x: 0, y: 0, z: 0.7}
-  mirrorNormal: {x: 0, y: 0, z: 1}
-  handPrefab: {fileID: 7860159785902570275, guid: c16e1c2363641754785f6357f9558525, type: 3}
-  headPrefab: {fileID: 1565915515836628941, guid: 5fba1076a99463a488221176302c9261, type: 3}
-  bodyPrefab: {fileID: 6447044102534172081, guid: 6171380971d754e47829a50deffa4403, type: 3}
-  particlePrefab: {fileID: 112561361}
-  collisionSound: {fileID: 8300000, guid: 89d1fcd2b23a3894fb3a294bda493116, type: 3}
---- !u!4 &1394094061
-Transform:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1394094059}
-  serializedVersion: 2
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0, y: 0.47, z: 2.6}
-  m_LocalScale: {x: 1, y: 1, z: 1}
-  m_ConstrainProportionsScale: 0
-  m_Children: []
-  m_Father: {fileID: 0}
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1001 &1425129741
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -1376,7 +1325,6 @@ SceneRoots:
   - {fileID: 948237271}
   - {fileID: 372890007}
   - {fileID: 2059071261}
-  - {fileID: 1394094061}
   - {fileID: 275216180}
   - {fileID: 1158395986}
   - {fileID: 1425129741}
