+++ CaveSoSe2025/Assets/Scripts/BodyCollision.cs
@@ -2,6 +2,9 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
+/// <summary>
+/// Detects collisions of the bodyparts and creates a particle system and sound if hands of avatar collide with each other.
+/// </summary>
 public class BodyCollision : MonoBehaviour
 {
     private ParticleSystem particleEffectPrefab;
@@ -18,7 +21,6 @@ public class BodyCollision : MonoBehaviour
 
     void OnCollisionEnter(Collision collision)
     {
-        // Debug.Log($"{gameObject.name} hat {collision.gameObject.name} berührt");
         if (gameObject.tag == "Hand" && collision.gameObject.tag == "Hand"){
            if (collision.contacts.Length > 0)
             {
@@ -40,31 +42,8 @@ public class BodyCollision : MonoBehaviour
         ParticleSystem instance = Instantiate(particleEffectPrefab, position, Quaternion.identity);
         instance.Play();
         Destroy(instance.gameObject, instance.main.startLifetime.constant + 0.2f);
-
-        //Variante mit, wenn eigenes Partikelsystem erstellt wird --> besser von außen zu übergeben, mehr Designfreiheit
-        /*GameObject psObj = new GameObject("CollisionParticles");
-        psObj.transform.position = position;
-
-        ParticleSystem ps = psObj.AddComponent<ParticleSystem>();
-
-        var main = ps.main;
-        main.startColor = Color.blue;
-        main.startSize = 0.1f;
-        main.startLifetime = 0.5f;
-        main.duration = 0.8f;
-        main.loop = false;
-
-        var emission = ps.emission;
-        emission.rateOverTime = 0f;
-        emission.SetBursts(new ParticleSystem.Burst[] {
-            new ParticleSystem.Burst(0f, 30)
-        });
-
-        ps.Play();
-        //particleSystem.Play();
-
-        Destroy(psObj, main.startLifetime.constant + 0.1f);*/
     }
+
     void PlayCollisionSound(Vector3 position)
     {
         if (collisionSound == null) return;
