--- CaveSoSe2025/Assets/Scripts/BodyPartsPositions.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7ee9164651a34744396756161122ae35
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
