--- CaveSoSe2025/Assets/Scripts/MovementEchoMultiple2.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 62a6fea5127cec048a6be9fc2b5b0fce
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
