--- CaveSoSe2025/Assets/Scripts/MovementRecord.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e70b41ee8e5097b40aca5b1c8dc6af96
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
