--- CaveSoSe2025/Assets/Scripts/PositionTransfer.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 342e7afcbea331d4eb1a5836c584ad07
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
